feat: Added delete call

+++ Backend/Controllers/VideoController.cs
@@ -29,6 +29,33 @@ public class VideoController : ControllerBase{
 
 
 
+    [HttpPut]
+    public IActionResult CreateVerdict(int id){
+        path = video.path;
+        // Call on FASTAPI, with path to video
+        // Await response
+        // Save JSONRESPONSE in database
+        // GET(response)
+        analyze.add(response)
+        return Success
+    }
+
+    def make_verdict(path):
+        return "{verdict: bad}"
+
+    def process_video(video):
+        // ....
+        // ...
+        retrun response
+
+    def API_CALL(path):
+        verdict = make_verdict(path)
+        process_video()
+        retrun Response(header, verdict)
+
+
+    localhost/FastApi/PostVerdict{string path}
+        ->      retrun JSONRESPONSE {"verdict": "bad"}
 
 
     [HttpPost]
@@ -43,18 +70,19 @@ public class VideoController : ControllerBase{
         return CreatedAtAction(nameof(Get), new { id = returnedVideo.Id }, returnedVideo);
     }
 
-    [HttpPut("{id}")]
-    public IActionResult Update(int id, Video video){
-        if (id != video.Id){
-            return BadRequest();
-        }
-        Video? prevVid = VideoService.Get(id);
-        if (prevVid is null){
-            return NotFound();
-        }
-        VideoService.Update(video);
-        return NoContent();
-    }
+// Imagine we wont need to PUT a video, atlest not for now
+//    [HttpPut("{id}")]
+//    public IActionResult Update(int id, Video video){
+//        if (id != video.Id){
+//            return BadRequest();
+//        }
+//        Video? prevVid = VideoService.Get(id);
+//        if (prevVid is null){
+//            return NotFound();
+//        }
+//        VideoService.Update(video);
+//        return NoContent();
+//    }
 
     [HttpDelete("{id}")]
     public IActionResult Delete(int id){
@@ -62,7 +90,11 @@ public class VideoController : ControllerBase{
         if (video is null){
             return NotFound();
         }
-        VideoService.Delete(id);
-        return NoContent();
+        if (System.IO.File.Exists(video.Path)){
+            System.IO.File.Delete(video.Path);
+            VideoService.Delete(id);
+            return NoContent();
+        }
+        return StatusCode(500);
     }
 }
